requirements.txt
discord.py
asyncpg
coloredlogs